Refactor ipc_os_prov_consumer.c and ipc_os_prov_producer.c
Refactor ipc_os_prov_consumer.c and ipc_os_prov_producer.c.
Split the code into functions, move them to the common directory
and prepare them to be reused in a future similar tests.
